+++ src/board/system76/common/peci.c
@@ -69,7 +69,7 @@ void peci_init(void) {
 
 // Returns positive completion code on success, negative completion code or
 // negative (0x1000 | status register) on PECI hardware error
-int peci_wr_pkg_config(uint8_t index, uint16_t param, uint32_t data) {
+int16_t peci_wr_pkg_config(uint8_t index, uint16_t param, uint32_t data) {
     // Wait for completion
     while (HOSTAR & 1) {}
     // Clear status
@@ -105,9 +105,9 @@ int peci_wr_pkg_config(uint8_t index, uint16_t param, uint32_t data) {
     // Wait for completion
     while (HOSTAR & 1) {}
 
-    int status = (int)HOSTAR;
+    int16_t status = (int16_t)HOSTAR;
     if (status & BIT(1)) {
-        int cc = (int)HORDDR;
+        int16_t cc = (int16_t)HORDDR;
         if (cc & 0x80) {
             return -cc;
         } else {
